Split spin.hpp into spin.hpp and spin.cpp to avoid odr violations
